fixed -Wall warnings

+++ src/TextWindow.cpp
@@ -2,10 +2,10 @@
 #include "string.h"
 
 void TextWindow::addText(const char* text){
-    int textSize=strlen(text);
+    unsigned int textSize=strlen(text);
     unsigned int startPos = winDimm.sizeX/2 - textSize/2;
     for(unsigned int pos=0; pos<textSize; pos++){
-        if(pos+2==winDimm.sizeX-2)
+        if(pos+2==(unsigned)winDimm.sizeX-2)
             return;
         drawingMarix[3][pos+startPos]=text[pos];
     }
